Revert "feat: OO Request testing scripts (#3)"
This reverts commit 6bcac43daa88978b1b953e16a273f7155503f2ae.
