+++ .gitignore
@@ -108,5 +108,3 @@ node_modules
 #Hardhat files
 cache
 artifacts
-
-.secret
