+++ hardhat.config.js
@@ -1,8 +1,5 @@
 require("@nomiclabs/hardhat-waffle");
 require("@nomiclabs/hardhat-ethers");
-const fs = require("fs");
-
-const privateKey = fs.readFileSync(".secret").toString().trim();
 
 // This is a sample Hardhat task. To learn how to create your own go to
 // https://hardhat.org/guides/create-task.html
@@ -55,23 +52,6 @@ module.exports = {
       hardfork: "london",
       gasPrice: 100,
       initialBaseFeePerGas: 1_000_000_000,
-    },
-    matic: {
-      /*      
-  New unified URL (recommended): https://polygon-rpc.com/
-  Use this new aggregated RPC link to avoid RPC disconnects. You can read more details on this here and here.
-
-  Use these alternative RPCs only if the above RPC fails:
-  https://rpc-mainnet.matic.network or
-  https://matic-mainnet.chainstacklabs.com or
-  https://rpc-mainnet.maticvigil.com or
-  https://rpc-mainnet.matic.quiknode.pro or
-  https://matic-mainnet-full-rpc.bwarelabs.com
-*/
-      url: "https://polygon-rpc.com/",
-      accounts: [privateKey],
-      // accountsBalance: "10000000000000000000000",
-      // gasPrice: 8000000000, // default is 'auto' which breaks chains without the london hardfork
     },
   },
 };
