+++ scripts/proposePriceToOO.js
@@ -18,14 +18,13 @@ async function main() {
   // const timeToAdvance = 5184000;
   const hardHatID = "31337";
 
-  const foundationAccount = "0x7a3A1c2De64f20EB5e916F40D11B01C441b2A8Dc";
+  const foundationAccount = "0x7a3a1c2de64f20eb5e916f40d11b01c441b2a8dc";
 
   const OOAddress = "0xc43767f4592df265b4a9f1a398b97ff24f38c6a6";
   const LSP_ADDRESS = "0x57C891D01605d456bBEa535c8E56EaAc4E2DFE11";
   const identifier =
     "0x7544414f5f4b50495f554d410000000000000000000000000000000000000000";
-  const ancData =
-    "0x737461727454696d657374616d703a313633333034363430302c6d617842617365496e746567726174696f6e733a32302c6d6178426f6e7573496e746567726174696f6e733a352c626f6e75734d696e56616c75653a353030303030302c626f6e7573496e746567726174696f6e734d756c7469706c6965723a322e30302c666c6f6f72496e746567726174696f6e733a30";
+  const ancData = "0x737461727454696d657374616d703a313633333034363430302c6d617842617365496e746567726174696f6e733a32302c6d6178426f6e7573496e746567726174696f6e733a352c626f6e75734d696e56616c75653a353030303030302c626f6e7573496e746567726174696f6e734d756c7469706c6965723a322e30302c666c6f6f72496e746567726174696f6e733a30"
   const timestamp = "1640872800";
   // LSP data.
   // const OOAddress = "0xc43767f4592df265b4a9f1a398b97ff24f38c6a6";
