+++ scripts/seedUmaToAccounts.js
@@ -1,0 +1,62 @@
+const hre = require("hardhat");
+const ethers = hre.ethers;
+const createERC20ContractInstance = require("./helpers/createERC20ContractInstance");
+
+async function main() {
+  const foundationAccount = "0x7a3a1c2de64f20eb5e916f40d11b01c441b2a8dc";
+
+  const hardhatAccountZero = "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266";
+  const hardhatAccountOne = "0x70997970c51812dc3a010c7d01b50e0d17dc79c8";
+  const hardhatAccountTwo = "0x3c44cdddb6a900fa2b585dd299e03d12fa4293bc";
+
+  try {
+    await hre.network.provider.request({
+      method: "hardhat_impersonateAccount",
+      params: [foundationAccount],
+    });
+
+    const signer = await ethers.provider.getSigner(foundationAccount);
+
+    const erc20 = await createERC20ContractInstance(signer, "1");
+    await hre.network.provider.send("hardhat_setBalance", [
+      foundationAccount,
+      "0x10000000000000000000000",
+    ]);
+
+    const txOne = await erc20.transfer(
+      hardhatAccountZero,
+      ethers.utils.parseEther("20")
+    );
+
+    const txTwo = await erc20.transfer(
+      hardhatAccountOne,
+      ethers.utils.parseEther("20")
+    );
+
+    const txThree = await erc20.transfer(
+      hardhatAccountTwo,
+      ethers.utils.parseEther("20")
+    );
+
+    console.log("TXOne", txOne);
+    console.log("TXTwo", txTwo);
+    console.log("TXThree", txThree);
+    // const txFour = await erc20.transfer()
+
+    await hre.network.provider.request({
+      method: "hardhat_stopImpersonatingAccount",
+      params: [foundationAccount],
+    });
+  } catch (err) {
+    console.log("err", err);
+  }
+}
+
+// We recommend this pattern to be able to use async/await everywhere
+// and properly handle errors.
+main()
+  .then(() => process.exit(0))
+  .catch((error) => {
+    console.error(error);
+    process.exit(1);
+  });
